Fixing & cleaning tests.
The build should now pass.

Testing can be done by:
$ ./dock
$ nose2

Tests should be sane now. So nobody should commit anything if it breaks any test. It would be great that every change to code is backed up by a new test that previously failed. With this we can guarantee that we will not be breaking each other's code.